fix: настроить поиск фильмов корректно
