+++ src/components/Footer/Footer.css
@@ -2,6 +2,7 @@
   max-width: 100%;
   background-color: #202020;
   font-style: normal;
+  margin-top: auto;
   padding: 0 70px;
   display: flex;
   flex-direction: column;
