+++ src/components/Movies/Movies.js
@@ -9,30 +9,14 @@ import filterMovies from "../../utils/moviesFilter";
 import filterShortMovies from "../../utils/shortMoviesFilter";
 
 function Movies(props) {
-  const [movies, setMovies] = useState([]);
+  const { onMoviesRequest, onMovieType, onSave, onDelete, isShortMovie, movies, inputError } = props;
 
-  useEffect(() => {
-    moviesApi.getBeatfilmMovies()
-      .then((res) => {
-        const filtredMovies = filterMovies(res, props.requestParameter);
-
-        if(props.isShortMovie) {
-          const shortMovies = filterShortMovies(filtredMovies);
-          setMovies(shortMovies);
-        } else {
-          setMovies(filtredMovies);
-        }
-      })
-      .catch(err => console.log('Ошибка', err));
-  },[props.requestParameter, props.isShortMovie]);
-
-  console.log(movies)
   return (
     <section className="movies">
       <div className="movies__container">
-        <SearchForm onMoviesRequest={props.onMoviesRequest} onMovieType={props.onMovieType} movieType={props.isShortMovie}/>
+        <SearchForm onMoviesRequest={onMoviesRequest} onMovieType={onMovieType} movieType={isShortMovie}/>
         <div className="movies__break"></div>
-        <MoviesCardList movies={movies} onSave={props.onSave} onDelete={props.onDelete}/>
+        <MoviesCardList movies={movies} onSave={onSave} onDelete={onDelete} inputError={inputError}/>
       </div>
     </section>
   )
