+++ src/components/MoviesCardList/MoviesCardList.css
@@ -38,6 +38,14 @@
   visibility: hidden;
 }
 
+.card-list__caption {
+  margin: 0;
+  font-style: normal;
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 20px;
+}
+
 @media (max-width: 770px) {
   .card-list {
     padding: 80px 0;
@@ -45,6 +53,12 @@
 
   .card-list__btn {
     margin-top: 60px;
+  }
+}
+
+@media (max-width: 700px) {
+  .card-list__caption {
+    font-size: 14px;
   }
 }
